+++ libraries/monitor/test/monitor_test.js
@@ -8,6 +8,7 @@ const MonitorManager = require('../src/monitormanager');
 
 suite(testing.suiteName(), function() {
   let monitorManager, monitor;
+  let errorBucket = [];
 
   suiteSetup(function() {
     monitorManager = new MonitorManager();
@@ -24,10 +25,15 @@ suite(testing.suiteName(), function() {
         allowExit: true,
       },
       verify: true,
+      errorConfig: {
+        reporter: 'TestReporter',
+        bucket: errorBucket,
+      },
     });
   });
 
   teardown(function() {
+    errorBucket.splice(0);
     monitorManager.reset();
     mockFs.restore();
   });
@@ -119,6 +125,7 @@ suite(testing.suiteName(), function() {
       assert(monitorManager.messages[0].Fields.duration);
       assert.equal(monitorManager.messages[0].Fields.status, 'success');
       assert(!monitorManager.messages[0].Fields.error);
+      assert.equal(errorBucket.length, 0);
     });
 
     test('unsuccessful async function', async function() {
@@ -130,6 +137,7 @@ suite(testing.suiteName(), function() {
       assert.equal(monitorManager.messages[1].Fields.name, 'expire');
       assert(monitorManager.messages[1].Fields.duration);
       assert.equal(monitorManager.messages[1].Fields.status, 'exception');
+      assert.equal(errorBucket.length, 1);
     });
 
     test('missing name', async function() {
@@ -138,6 +146,7 @@ suite(testing.suiteName(), function() {
       assert.equal(monitorManager.messages.length, 2);
       assert.equal(monitorManager.messages[0].Type, 'monitor.error');
       assert.equal(monitorManager.messages[0].Fields.code, 'ERR_ASSERTION');
+      assert.equal(errorBucket.length, 2); // One for normal error and one for missing name
     });
   });
 
@@ -260,7 +269,8 @@ suite(testing.suiteName(), function() {
       ], (done, code, output) => {
         assert.equal(code, 1);
         assert(output.includes('Error: hello there'));
-        assert(JSON.parse(output));
+        assert(JSON.parse(output.split('\n')[0]));
+        assert.equal(output.split('\n')[1], 'REPORTED ERROR');
         done();
       });
     });
@@ -284,7 +294,8 @@ suite(testing.suiteName(), function() {
       ], (done, code, output) => {
         assert.equal(code, 1);
         assert(output.includes('whaaa'));
-        assert(JSON.parse(output));
+        assert(JSON.parse(output.split('\n')[0]));
+        assert.equal(output.split('\n')[1], 'REPORTED ERROR');
         done();
       });
     });