+++ services/web-server/src/login/strategies/github.js
@@ -4,10 +4,11 @@ const passport = require('passport');
 const { Strategy } = require('passport-github');
 const taskcluster = require('taskcluster-client');
 const User = require('../User');
-const identityFromClientId = require('../../utils/identityFromClientId');
-const { encode, decode } = require('../../utils/codec');
 const login = require('../../utils/login');
 const WebServerError = require('../../utils/WebServerError');
+const tryCatch = require('../../utils/tryCatch');
+const { encode, decode } = require('../../utils/codec');
+const GithubClient = require('../clients/GithubClient');
 
 const debug = Debug('strategies.github');
 
@@ -20,19 +21,29 @@ module.exports = class Github {
 
     Object.assign(this, strategyCfg);
 
-    this.rootUrl = cfg.taskcluster.rootUrl;
     this.identityProviderId = 'github';
+    this.githubClient = new GithubClient();
   }
 
-  async getUser({ userId }) {
+  async getUser({ username, userId }) {
     const user = new User();
+    const [githubErr, githubUser] = await tryCatch(this.githubClient.userFromUsername(username));
 
-    user.identity = `${this.identityProviderId}/${encode(userId)}`;
+    // 404 means the user doesn't exist; otherwise, throw the error up the chain
+    if (githubErr) {
+      if (githubErr.status === 404) {
+        return;
+      }
+      throw githubErr;
+    }
 
-    if (!user.identity) {
-      debug('No recognized identity providers');
+    if (githubUser.id !== userId) {
+      debug(`Github user id ${githubUser.id} does not match userId ${userId} from the identity.`);
+
       return;
     }
+
+    user.identity = `${this.identityProviderId}/${userId}|${encode(username)}`;
 
     // take a user and attach roles to it
     // this.addRoles(userProfile, user);
@@ -41,20 +52,9 @@ module.exports = class Github {
   }
 
   userFromIdentity(identity) {
-    const encodedUserId = identity.split('/')[1];
-    const userId = decode(encodedUserId);
+    const [userId, username = ''] = identity.split('/')[1].split('|');
 
-    return this.getUser({ userId });
-  }
-
-  userFromClientId(clientId) {
-    const identity = identityFromClientId(clientId);
-
-    if (!identity) {
-      return;
-    }
-
-    return this.userFromIdentity(identity);
+    return this.getUser({ username: decode(username), userId: Number(userId) });
   }
 
   useStrategy(app, cfg) {
@@ -84,7 +84,7 @@ module.exports = class Github {
           callbackURL: `${cfg.app.publicUrl}${callback}`,
         },
         async (accessToken, refreshToken, profile, next) => {
-          const user = await this.getUser({ userId: profile.username });
+          const user = await this.getUser({ username: profile.username, userId: Number(profile.id) });
 
           if (!user) {
             // Don't report much to the user, to avoid revealing sensitive information, although