+++ services/worker-manager/test/provider_google_test.js
@@ -33,6 +33,7 @@ helper.secrets.mockSuite(testing.suiteName(), ['taskcluster', 'azure'], function
         project: 'testy',
         instancePermissions: [],
         creds: '{}',
+        workerServiceAccountId: '12345',
       },
     });
     workerPool = await helper.WorkerPool.create({
@@ -48,7 +49,7 @@ helper.secrets.mockSuite(testing.suiteName(), ['taskcluster', 'azure'], function
         capacityPerInstance: 1,
         machineType: 'n1-standard-2',
         regions: ['us-east1'],
-        userData: {},
+        workerConfig: {},
         scheduling: {},
         networkInterfaces: [],
         disks: [],