+++ ui/src/utils/constants.js
@@ -232,28 +232,29 @@ export const GROUP_NOTIFY_SUCCESS_KEY = 'group-notify-success';
 
 // Worker Manager constants
 export const NULL_PROVIDER = 'null-provider';
-export const GCP = 'GCP';
-export const PROVIDERS = new Map([[`${GCP}`, 'google']]);
-export const PROVIDER_CONFIGS = new Map([
+export const PROVIDER_DEFAULT_CONFIGS = new Map([
+  // providerType : default config
   [
-    `${GCP}`,
+    'google',
     {
       minCapacity: 0,
       maxCapacity: 0,
       capacityPerInstance: 1,
       machineType: 'n1-highcpu-8',
       regions: ['us-west2'],
-      userData: {},
+      workerConfig: {},
       scheduling: {},
       networkInterfaces: [{}],
       disks: [{}],
     },
   ],
+  ['static', {}],
 ]);
 export const NULL_WORKER_POOL = {
   workerPoolId: '/',
+  providerId: '',
   description: '',
   owner: '',
   emailOnError: false,
-  config: PROVIDER_CONFIGS.get(GCP),
+  config: {},
 };