+++ ui/src/utils/notify.js
@@ -4,7 +4,7 @@ import logo from '../images/brandLogo.png';
  Display a Desktop notification using the Notifications API to the user.
  */
 export default ({ body, icon }) => {
-  const notification = new Notification(process.env.APPLICATION_NAME, {
+  const notification = new Notification(window.env.APPLICATION_NAME, {
     icon: icon || logo,
     body,
   });